# Send images to Imaginary docker to generate previews