--- a/lib/private/Http/Client/Client.php
+++ b/lib/private/Http/Client/Client.php
@@ -292,7 +292,8 @@ class Client implements IClient {
 			unset($options['body']);
 		}
 		$response = $this->client->request('post', $uri, $this->buildRequestOptions($options));
-		return new Response($response);
+		$isStream = isset($options['stream']) && $options['stream'];
+		return new Response($response, $isStream);
 	}
 
 	/**