… lib/composer/composer/autoload_classmap.php
… lib/composer/composer/autoload_static.php