@PVince81 2020-11-16
Quick benchmark with 4 jpg files with roughly the same size:
- with Imaginary: ~4.7 second
- native way: ~7.5 seconds