@rullzer 2020-11-17
@PVince81 why not do this as an app?