@rullzer 2020-11-17
> why is the maxX/maxY always 4096 even though the UI is requesting less than 400x400 ? (did I find a bug?)

No that was the part I also ran into when I wrote an imaginary app to play preview provider.
The preview provider is only used to generate the initial preview (so the highest resolution one)
Then we use GD to create the downsized images.