@PVince81 2020-11-17
> @PVince81 why not do this as an app?

because only a prototype for benchmarking, it's easier to push a PR that way.
in any case this is to be discussed first.
if worth it, we can make an app, yes.

I suspect we might need to tweak the preview manager to make it possible to override the default preview providers, which somehow didn't work when I tested (hence the early "return" in this PR)