@PVince81 2020-11-17
- [ ] should we use Imaginary also for the small resizes from the cached image instead of GD ? (make the preview manager support such operations by third parties ?)