@PVince81 2020-11-17
work to be continued in this app: https://github.com/rullzer/imaginary/issues/1

I'll leave the PR open until we have transferred the tasks and code