@CarlSchwan 2022-02-11
Benchmarks with s3:

Small images (256x256):

* imaginary: ~0.7s
* current Image provider: ~0.4s

Bigger image: HD

* imaginary: ~1.5s
* current image providers: ~2.5s

So it seems that the overhead from the network is really only worth it for larger images there the multithreading of imaginary really helps. This should be most pictures as taking small 256x356 images is not really a thing anymore