@CarlSchwan 2022-02-11
Found an issue, we don't get the width and height of the generated preview, so we currently assume it's 4096x4096 and this is causing an issue for small images (< 1000x1000) as we then generate an additional preview for nothing.

I made a PR upstream https://github.com/h2non/imaginary/pull/382 to solve this