@CarlSchwan 2022-02-15
> Code looks fine to me. Can't approve as I'm co-author 😄
> 
> Did you test that the default config still uses the old preview provider ?

Yes, I tested it when doing the benchmarks.