@juliushaertl 2022-03-17
Fixed the missing phpdoc on `IStreamImage`, CI should be happy now